+++ core/governance/engine_update_market_test.go
@@ -14,13 +14,16 @@ package governance_test
 
 import (
 	"context"
+	"fmt"
 	"testing"
 	"time"
 
 	"code.vegaprotocol.io/vega/core/governance"
 	"code.vegaprotocol.io/vega/core/netparams"
 	"code.vegaprotocol.io/vega/core/types"
+	"code.vegaprotocol.io/vega/libs/num"
 	vgrand "code.vegaprotocol.io/vega/libs/rand"
+	vegapb "code.vegaprotocol.io/vega/protos/vega"
 	datapb "code.vegaprotocol.io/vega/protos/vega/data/v1"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
@@ -28,10 +31,16 @@ import (
 
 func TestProposalForMarketUpdate(t *testing.T) {
 	t.Run("Submitting a proposal for market update succeeds", testSubmittingProposalForMarketUpdateSucceeds)
+	t.Run("Submitting a proposal for market update with internal time termination succeeds", testSubmittingProposalForMarketUpdateWithInternalTimeTerminationSucceeds)
+	t.Run("Submitting a proposal for market update with internal settling fails", testSubmittingProposalForMarketUpdateWithInternalTimeSellingFails)
+	t.Run("Submitting a proposal for market update with external termination using internal time key succeeds", testSubmittingProposalForMarketUpdateWithExternalSourceUsingInternalKeyTimeForTerminationSucceeds)
 	t.Run("Submitting a proposal for market update on unknown market fails", testSubmittingProposalForMarketUpdateForUnknownMarketFails)
+	t.Run("Submitting a proposal with internal time termination for market update on unknown market fails", testSubmittingProposalForMarketUpdateWithInternalTimeTerminationForUnknownMarketFails)
+
 	t.Run("Submitting a proposal for market update for not-enacted market fails", testSubmittingProposalForMarketUpdateForNotEnactedMarketFails)
 	t.Run("Submitting a proposal for market update with insufficient equity-like share fails", testSubmittingProposalForMarketUpdateWithInsufficientEquityLikeShareFails)
 	t.Run("Pre-enactment of market update proposal succeeds", testPreEnactmentOfMarketUpdateSucceeds)
+	t.Run("Pre-enactment of market with internal time termination update proposal succeeds", testPreEnactmentOfMarketUpdateWithInternalTimeTerminationSucceeds)
 
 	t.Run("Rejecting a proposal for market update succeeds", testRejectingProposalForMarketUpdateSucceeds)
 
@@ -48,7 +57,152 @@ func testSubmittingProposalForMarketUpdateSucceeds(t *testing.T) {
 
 	// given
 	proposer := vgrand.RandomStr(5)
-	proposal := eng.newProposalForMarketUpdate("market-1", proposer, eng.tsvc.GetTimeNow(), nil, nil)
+	proposal := eng.newProposalForMarketUpdate("market-1", proposer, eng.tsvc.GetTimeNow(), nil, nil, true)
+	marketID := proposal.MarketUpdate().MarketID
+
+	// setup
+	eng.ensureTokenBalanceForParty(t, proposer, 1000)
+	eng.ensureEquityLikeShareForMarketAndParty(t, marketID, proposer, 0.1)
+	eng.ensureExistingMarket(t, marketID)
+
+	// expect
+	eng.expectOpenProposalEvent(t, proposer, proposal.ID)
+
+	// when
+	toSubmit, err := eng.submitProposal(t, proposal)
+
+	// then
+	require.NoError(t, err)
+	require.NotNil(t, toSubmit)
+}
+
+func testSubmittingProposalForMarketUpdateWithInternalTimeTerminationSucceeds(t *testing.T) {
+	eng := getTestEngine(t)
+	defer eng.ctrl.Finish()
+
+	// given
+	proposer := vgrand.RandomStr(5)
+	proposal := eng.newProposalForMarketUpdate("market-1", proposer, eng.tsvc.GetTimeNow(), nil, nil, false)
+	marketID := proposal.MarketUpdate().MarketID
+
+	// setup
+	eng.ensureTokenBalanceForParty(t, proposer, 1000)
+	eng.ensureEquityLikeShareForMarketAndParty(t, marketID, proposer, 0.1)
+	eng.ensureExistingMarket(t, marketID)
+
+	// expect
+	eng.expectOpenProposalEvent(t, proposer, proposal.ID)
+
+	// when
+	toSubmit, err := eng.submitProposal(t, proposal)
+
+	// then
+	require.NoError(t, err)
+	require.NotNil(t, toSubmit)
+}
+
+func testSubmittingProposalForMarketUpdateWithInternalTimeSellingFails(t *testing.T) {
+	eng := getTestEngine(t)
+	defer eng.ctrl.Finish()
+
+	// given
+	proposer := vgrand.RandomStr(5)
+
+	id := eng.newProposalID()
+	now := eng.tsvc.GetTimeNow()
+	tm := time.Now().Add(time.Hour * 24 * 365)
+	_, termBinding := produceTimeTriggeredDataSourceSpec(tm)
+	proposal := types.Proposal{
+		ID:        "market-1",
+		Reference: "ref-" + id,
+		Party:     proposer,
+		State:     types.ProposalStateOpen,
+		Terms: &types.ProposalTerms{
+			ClosingTimestamp:    now.Add(96 * time.Hour).Unix(),
+			EnactmentTimestamp:  now.Add(4 * 48 * time.Hour).Unix(),
+			ValidationTimestamp: now.Add(2 * time.Hour).Unix(),
+			Change: &types.ProposalTermsUpdateMarket{
+				UpdateMarket: &types.UpdateMarket{
+					MarketID: vgrand.RandomStr(5),
+					Changes: &types.UpdateMarketConfiguration{
+						Instrument: &types.UpdateInstrumentConfiguration{
+							Code: "CRYPTO:GBPVUSD/JUN20",
+							Product: &types.UpdateInstrumentConfigurationFuture{
+								Future: &types.UpdateFutureProduct{
+									QuoteName: "VUSD",
+									DataSourceSpecForSettlementData: *types.NewDataSourceDefinition(
+										vegapb.DataSourceDefinitionTypeExt,
+									).SetTimeTriggerConditionConfig(
+										[]*types.DataSourceSpecCondition{
+											{
+												Operator: datapb.Condition_OPERATOR_GREATER_THAN_OR_EQUAL,
+												Value:    fmt.Sprintf("%d", tm.UnixNano()),
+											},
+										},
+									),
+									DataSourceSpecForTradingTermination: *types.NewDataSourceDefinition(
+										vegapb.DataSourceDefinitionTypeExt,
+									).SetTimeTriggerConditionConfig(
+										[]*types.DataSourceSpecCondition{
+											{
+												Operator: datapb.Condition_OPERATOR_GREATER_THAN_OR_EQUAL,
+												Value:    fmt.Sprintf("%d", tm.UnixNano()),
+											},
+										},
+									),
+									DataSourceSpecBinding: termBinding,
+								},
+							},
+						},
+						RiskParameters: &types.UpdateMarketConfigurationLogNormal{
+							LogNormal: &types.LogNormalRiskModel{
+								RiskAversionParameter: num.DecimalFromFloat(0.01),
+								Tau:                   num.DecimalFromFloat(0.00011407711613050422),
+								Params: &types.LogNormalModelParams{
+									Mu:    num.DecimalZero(),
+									R:     num.DecimalFromFloat(0.016),
+									Sigma: num.DecimalFromFloat(0.09),
+								},
+							},
+						},
+						Metadata:                []string{"asset_class:fx/crypto", "product:futures"},
+						LpPriceRange:            num.DecimalFromFloat(0.95),
+						LinearSlippageFactor:    num.DecimalFromFloat(0.1),
+						QuadraticSlippageFactor: num.DecimalFromFloat(0.1),
+					},
+				},
+			},
+		},
+		Rationale: &types.ProposalRationale{
+			Description: "some description",
+		},
+	}
+
+	marketID := proposal.MarketUpdate().MarketID
+	// setup
+	eng.ensureTokenBalanceForParty(t, proposer, 1000)
+	eng.ensureEquityLikeShareForMarketAndParty(t, marketID, proposer, 0.1)
+	eng.ensureExistingMarket(t, marketID)
+
+	// expect
+	eng.expectRejectedProposalEvent(t, proposer, proposal.ID, types.ProposalErrorInvalidFutureProduct)
+
+	// when
+	toSubmit, err := eng.submitProposal(t, proposal)
+
+	// then
+	assert.Error(t, err, governance.ErrSettlementWithInternalDataSourceIsNotAllowed)
+	require.Nil(t, toSubmit)
+}
+
+func testSubmittingProposalForMarketUpdateWithExternalSourceUsingInternalKeyTimeForTerminationSucceeds(t *testing.T) {
+	eng := getTestEngine(t)
+	defer eng.ctrl.Finish()
+
+	// given
+	proposer := vgrand.RandomStr(5)
+	filter, binding := produceTimeTriggeredDataSourceSpec(time.Now())
+	proposal := eng.newProposalForMarketUpdate("market-1", proposer, eng.tsvc.GetTimeNow(), filter, binding, true)
 	marketID := proposal.MarketUpdate().MarketID
 
 	// setup
@@ -73,7 +227,7 @@ func TestSubmittingProposalForMarketUpdateWithEarlyTerminationSucceeds(t *testin
 
 	// given
 	proposer := vgrand.RandomStr(5)
-	proposal := eng.newProposalForMarketUpdate("market-1", proposer, eng.tsvc.GetTimeNow(), nil, nil)
+	proposal := eng.newProposalForMarketUpdate("market-1", proposer, eng.tsvc.GetTimeNow(), nil, nil, true)
 	marketID := proposal.MarketUpdate().MarketID
 
 	proposal.Terms.Change.(*types.ProposalTermsUpdateMarket).UpdateMarket.Changes.Instrument.Product.(*types.UpdateInstrumentConfigurationFuture).Future.DataSourceSpecForTradingTermination.UpdateFilters(
@@ -116,7 +270,31 @@ func testSubmittingProposalForMarketUpdateForUnknownMarketFails(t *testing.T) {
 
 	// given
 	proposer := vgrand.RandomStr(5)
-	proposal := eng.newProposalForMarketUpdate("״market-1", proposer, eng.tsvc.GetTimeNow(), nil, nil)
+	proposal := eng.newProposalForMarketUpdate("״market-1", proposer, eng.tsvc.GetTimeNow(), nil, nil, true)
+	marketID := proposal.MarketUpdate().MarketID
+
+	// setup
+	eng.ensureTokenBalanceForParty(t, proposer, 123456789)
+	eng.ensureNonExistingMarket(t, marketID)
+
+	// expect
+	eng.expectRejectedProposalEvent(t, proposer, proposal.ID, types.ProposalErrorInvalidMarket)
+
+	// when
+	toSubmit, err := eng.submitProposal(t, proposal)
+
+	// then
+	require.ErrorIs(t, governance.ErrMarketDoesNotExist, err)
+	require.Nil(t, toSubmit)
+}
+
+func testSubmittingProposalForMarketUpdateWithInternalTimeTerminationForUnknownMarketFails(t *testing.T) {
+	eng := getTestEngine(t)
+	defer eng.ctrl.Finish()
+
+	// given
+	proposer := vgrand.RandomStr(5)
+	proposal := eng.newProposalForMarketUpdate("״market-1", proposer, eng.tsvc.GetTimeNow(), nil, nil, false)
 	marketID := proposal.MarketUpdate().MarketID
 
 	// setup
@@ -140,7 +318,7 @@ func testSubmittingProposalForMarketUpdateForNotEnactedMarketFails(t *testing.T)
 
 	// given
 	proposer := vgrand.RandomStr(5)
-	newMarketProposal := eng.newProposalForNewMarket(proposer, eng.tsvc.GetTimeNow(), nil, nil)
+	newMarketProposal := eng.newProposalForNewMarket(proposer, eng.tsvc.GetTimeNow(), nil, nil, true)
 	marketID := newMarketProposal.ID
 
 	// setup
@@ -157,7 +335,7 @@ func testSubmittingProposalForMarketUpdateForNotEnactedMarketFails(t *testing.T)
 	assert.True(t, toSubmit.IsNewMarket())
 
 	// given
-	updateMarketProposal := eng.newProposalForMarketUpdate("״market-1", proposer, eng.tsvc.GetTimeNow(), nil, nil)
+	updateMarketProposal := eng.newProposalForMarketUpdate("״market-1", proposer, eng.tsvc.GetTimeNow(), nil, nil, true)
 	updateMarketProposal.MarketUpdate().MarketID = marketID
 
 	// setup
@@ -181,7 +359,7 @@ func testSubmittingProposalForMarketUpdateWithInsufficientEquityLikeShareFails(t
 
 	// given
 	party := vgrand.RandomStr(5)
-	proposal := eng.newProposalForMarketUpdate("״market-1", party, eng.tsvc.GetTimeNow(), nil, nil)
+	proposal := eng.newProposalForMarketUpdate("״market-1", party, eng.tsvc.GetTimeNow(), nil, nil, true)
 	marketID := proposal.MarketUpdate().MarketID
 
 	// setup
@@ -208,7 +386,123 @@ func testPreEnactmentOfMarketUpdateSucceeds(t *testing.T) {
 	// Submit proposal.
 	// given
 	proposer := vgrand.RandomStr(5)
-	proposal := eng.newProposalForMarketUpdate("״market-1", proposer, eng.tsvc.GetTimeNow(), nil, nil)
+	proposal := eng.newProposalForMarketUpdate("״market-1", proposer, eng.tsvc.GetTimeNow(), nil, nil, true)
+	marketID := proposal.MarketUpdate().MarketID
+
+	// setup
+	eng.ensureEquityLikeShareForMarketAndParty(t, marketID, proposer, 0.7)
+	eng.ensureExistingMarket(t, marketID)
+	eng.ensureTokenBalanceForParty(t, proposer, 1)
+	eng.ensureAllAssetEnabled(t)
+
+	// expect
+	eng.expectOpenProposalEvent(t, proposer, proposal.ID)
+
+	// when
+	_, err := eng.submitProposal(t, proposal)
+
+	// then
+	require.NoError(t, err)
+
+	// Vote 'YES' with 10 tokens.
+	// given
+	voterWithToken1 := vgrand.RandomStr(5)
+
+	// setup
+	eng.ensureTokenBalanceForParty(t, voterWithToken1, 10)
+	eng.ensureEquityLikeShareForMarketAndParty(t, marketID, voterWithToken1, 0)
+
+	// expect
+	eng.expectVoteEvent(t, voterWithToken1, proposal.ID)
+
+	// when
+	err = eng.addYesVote(t, voterWithToken1, proposal.ID)
+
+	// then
+	require.NoError(t, err)
+
+	// Vote 'NO' with 2 tokens.
+	// given
+	voterWithToken2 := vgrand.RandomStr(5)
+
+	// setup
+	eng.ensureTokenBalanceForParty(t, voterWithToken2, 2)
+	eng.ensureEquityLikeShareForMarketAndParty(t, marketID, voterWithToken2, 0)
+
+	// expect
+	eng.expectVoteEvent(t, voterWithToken2, proposal.ID)
+
+	// then
+	err = eng.addNoVote(t, voterWithToken2, proposal.ID)
+
+	// then
+	require.NoError(t, err)
+
+	// Close the proposal.
+	// given
+	afterClosing := time.Unix(proposal.Terms.ClosingTimestamp, 0).Add(time.Second)
+
+	// setup
+	eng.ensureStakingAssetTotalSupply(t, 13)
+	eng.ensureTokenBalanceForParty(t, voterWithToken1, 10)
+	eng.ensureTokenBalanceForParty(t, voterWithToken2, 2)
+
+	// expect
+	eng.expectPassedProposalEvent(t, proposal.ID)
+	eng.expectVoteEvents(t)
+	eng.expectGetMarketState(t, marketID)
+
+	// when
+	eng.OnTick(context.Background(), afterClosing)
+
+	// Enact the proposal.
+	// given
+	afterEnactment := time.Unix(proposal.Terms.EnactmentTimestamp, 0).Add(time.Second)
+	existingMarket := types.Market{
+		ID: marketID,
+		TradableInstrument: &types.TradableInstrument{
+			Instrument: &types.Instrument{
+				Name: vgrand.RandomStr(10),
+				Product: &types.InstrumentFuture{
+					Future: &types.Future{
+						SettlementAsset: "BTC",
+					},
+				},
+			},
+		},
+		DecimalPlaces:         3,
+		PositionDecimalPlaces: 4,
+		OpeningAuction: &types.AuctionDuration{
+			Duration: 42,
+		},
+	}
+
+	// setup
+	eng.ensureGetMarket(t, marketID, existingMarket)
+
+	// when
+	enacted, _ := eng.OnTick(context.Background(), afterEnactment)
+
+	// then
+	require.NotEmpty(t, enacted)
+	require.True(t, enacted[0].IsUpdateMarket())
+	updatedMarket := enacted[0].UpdateMarket()
+	assert.Equal(t, existingMarket.ID, updatedMarket.ID)
+	assert.Equal(t, existingMarket.TradableInstrument.Instrument.Name, updatedMarket.TradableInstrument.Instrument.Name)
+	assert.Equal(t, existingMarket.TradableInstrument.Instrument.Product.(*types.InstrumentFuture).Future.SettlementAsset, updatedMarket.TradableInstrument.Instrument.Product.(*types.InstrumentFuture).Future.SettlementAsset)
+	assert.Equal(t, existingMarket.DecimalPlaces, updatedMarket.DecimalPlaces)
+	assert.Equal(t, existingMarket.PositionDecimalPlaces, updatedMarket.PositionDecimalPlaces)
+	assert.Equal(t, existingMarket.OpeningAuction.Duration, updatedMarket.OpeningAuction.Duration)
+}
+
+func testPreEnactmentOfMarketUpdateWithInternalTimeTerminationSucceeds(t *testing.T) {
+	eng := getTestEngine(t)
+	defer eng.ctrl.Finish()
+
+	// Submit proposal.
+	// given
+	proposer := vgrand.RandomStr(5)
+	proposal := eng.newProposalForMarketUpdate("״market-1", proposer, eng.tsvc.GetTimeNow(), nil, nil, false)
 	marketID := proposal.MarketUpdate().MarketID
 
 	// setup
@@ -323,7 +617,7 @@ func testRejectingProposalForMarketUpdateSucceeds(t *testing.T) {
 
 	// given
 	party := vgrand.RandomStr(5)
-	proposal := eng.newProposalForMarketUpdate("market-1", party, eng.tsvc.GetTimeNow(), nil, nil)
+	proposal := eng.newProposalForMarketUpdate("market-1", party, eng.tsvc.GetTimeNow(), nil, nil, true)
 	marketID := proposal.MarketUpdate().MarketID
 
 	// setup
@@ -367,7 +661,7 @@ func testVotingWithoutMinimumTokenHoldersAndEquityLikeShareMakesMarketUpdateProp
 	// Submit proposal.
 	// given
 	proposer := vgrand.RandomStr(5)
-	proposal := eng.newProposalForMarketUpdate("״market-1", proposer, eng.tsvc.GetTimeNow(), nil, nil)
+	proposal := eng.newProposalForMarketUpdate("״market-1", proposer, eng.tsvc.GetTimeNow(), nil, nil, true)
 	marketID := proposal.MarketUpdate().MarketID
 
 	// setup
@@ -448,7 +742,7 @@ func testVotingWithMajorityOfYesFromTokenHoldersMakesMarketUpdateProposalPassed(
 	// Submit proposal.
 	// given
 	proposer := vgrand.RandomStr(5)
-	proposal := eng.newProposalForMarketUpdate("״market-1", proposer, eng.tsvc.GetTimeNow(), nil, nil)
+	proposal := eng.newProposalForMarketUpdate("״market-1", proposer, eng.tsvc.GetTimeNow(), nil, nil, true)
 	marketID := proposal.MarketUpdate().MarketID
 
 	// setup
@@ -561,7 +855,7 @@ func testVotingWithMajorityOfNoFromTokenHoldersMakesMarketUpdateProposalDeclined
 	// Submit proposal.
 	// given
 	proposer := vgrand.RandomStr(5)
-	proposal := eng.newProposalForMarketUpdate("market-1", proposer, eng.tsvc.GetTimeNow(), nil, nil)
+	proposal := eng.newProposalForMarketUpdate("market-1", proposer, eng.tsvc.GetTimeNow(), nil, nil, true)
 	marketID := proposal.MarketUpdate().MarketID
 
 	// setup
@@ -676,7 +970,7 @@ func testVotingWithoutTokenAndMajorityOfYesFromEquityLikeShareHoldersMakesMarket
 	// Submit proposal.
 	// given
 	proposer := vgrand.RandomStr(5)
-	proposal := eng.newProposalForMarketUpdate("market-1", proposer, eng.tsvc.GetTimeNow(), nil, nil)
+	proposal := eng.newProposalForMarketUpdate("market-1", proposer, eng.tsvc.GetTimeNow(), nil, nil, true)
 	marketID := proposal.MarketUpdate().MarketID
 
 	// setup
@@ -777,7 +1071,7 @@ func testVotingWithoutTokenAndMajorityOfNoFromEquityLikeShareHoldersMakesMarketU
 	eng.ensureNetworkParameter(t, netparams.GovernanceProposalUpdateMarketRequiredParticipation, "0.5")
 
 	proposer := vgrand.RandomStr(5)
-	proposal := eng.newProposalForMarketUpdate("market-1", proposer, eng.tsvc.GetTimeNow(), nil, nil)
+	proposal := eng.newProposalForMarketUpdate("market-1", proposer, eng.tsvc.GetTimeNow(), nil, nil, true)
 	marketID := proposal.MarketUpdate().MarketID
 
 	// setup